Optimizing- Milestone.5

+++ app/src/main/java/com/tuwaiq/to_dolistapp/DashBoardActivity.kt
@@ -3,6 +3,7 @@ package com.tuwaiq.to_dolistapp
 import android.content.Context
 import android.content.DialogInterface
 import android.content.Intent
+import android.graphics.PaintFlagsDrawFilter
 import android.os.Bundle
 import android.service.controls.actions.FloatAction
 import android.view.LayoutInflater
@@ -114,8 +115,18 @@ class DashboardActivity : AppCompatActivity() {
                             activity.updateToDo(list[p1])
                         }
                         R.id.menu_delete->{
-                            activity.dbHandler.deleteToDo(list[p1].id)
-                            activity.refreshList()
+                            val dialog = AlertDialog.Builder(activity)
+                            dialog.setTitle("Are you sure?")
+                            dialog.setMessage("Do you want to delete this task?")
+                            dialog.setPositiveButton("Continue") {_: DialogInterface, _: Int ->
+                                activity.dbHandler.deleteToDo(list[p1].id)
+                                activity.refreshList()
+                            }
+                            dialog.setNegativeButton("Cancel") {_: DialogInterface, _: Int ->
+
+                            }
+                            dialog.show()
+
                         }
                         R.id.menu_mark_as_completed->{
                             activity.dbHandler.updateToDoItemCompletedStatus(list[p1].id,true)
