+++ app/src/main/java/com/tuwaiq/to_dolistapp/ItemActivity.kt
@@ -44,6 +44,7 @@ class ItemActivity : AppCompatActivity() {
 
         fabItem.setOnClickListener {
             val dialog = AlertDialog.Builder(this)
+            dialog.setTitle("Add ToDo item")
             val view = layoutInflater.inflate(R.layout.dialog_dashboard, null)
             val toDoName = view.findViewById<EditText>(R.id.tv_todo)
             dialog.setView(view)
@@ -66,6 +67,7 @@ class ItemActivity : AppCompatActivity() {
 
     fun updateItem(item: ToDoItem){
         val dialog = AlertDialog.Builder(this)
+        dialog.setTitle("Update ToDo item")
         val view = layoutInflater.inflate(R.layout.dialog_dashboard, null)
         val toDoName = view.findViewById<EditText>(R.id.tv_todo)
         toDoName.setText(item.itemName)
@@ -110,8 +112,18 @@ class ItemActivity : AppCompatActivity() {
                 activity.dbHandler.updateToDoItem(list[p1])
             }
             holder.delete.setOnClickListener{
-                activity.dbHandler.deleteToDoItem(list[p1].id)
-                activity.refreshList()
+                val dialog = AlertDialog.Builder(activity)
+                dialog.setTitle("Are you sure?")
+                dialog.setMessage("Do you want to delete this item?")
+                dialog.setPositiveButton("Continue") {_: DialogInterface, _: Int ->
+                    activity.dbHandler.deleteToDoItem(list[p1].id)
+                    activity.refreshList()
+                }
+                dialog.setNegativeButton("Cancel") {_: DialogInterface, _: Int ->
+
+                }
+                dialog.show()
+
             }
             holder.edit.setOnClickListener{
                 activity.updateItem(list[p1])
